Vibra: Remove vibrator AOSP implementation from libhardware_legacy
To enable the vendor implementation of the vibrator HAL, it is needed
to use hardware modules, a default one for AOSP and the vendor one.
The legacy implementation is no more needed so let's remove it.

This change is related to other changes in:
 - hardware/libhardware
 - frameworks/base

Change-Id: I844279f5535289f079d412fdc44c5cb3c9c1130c
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-10085
Category: aosp improvement
Domain: AudioComms-HwModules-Vibrator
Origin: upstream https://android-review.googlesource.com/50291
Upstream-Candidate: yes
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: https://android.intel.com:443/236132
Reviewed-on: https://android.intel.com:443/304016