[PORT FROM MAIN] [VIBRA] Remove vibrator AOSP implementation from libhardware_legacy
BZ: 49760 126632

To enable the vendor implementation of the vibrator HAL, it is needed
to use hardware modules, a default one for AOSP and the vendor one.
The implementation in libhardware_legacy is no more needed so remove
it from the file system.

Orig-Change-Id: I5817ef7543d65a991e593469b76d52b978dd1254
Category: aosp improvement
Domain: AudioComms-HwModules-Vibrator
Origin: internal
Upstream-Candidate: yes
Change-Id: Id64370972a7014812f05f061f9175e8405fcdd5f
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>