[PORT FROM MAIN] When setting a null device for an audio output, force the next routing request.
BZ: 53856

When a null device is set for an audio ouput, the outputDesc->mDevice is not updated,
thus the routing update is not done next time the device is selected for that output.
To meet the device change conditions, a boolean member mForceRouting has been
introduced to force the routing next time the output is set with a non-null device.

Also contains a fix made afterward in:
    [PORT FROM MAIN] init of all attributes&warning fix

    Change-Id: Ib81f9868efc2a8614ea4395f7722c09775d2b9d3
    Origin-Change-Id: I9ba288b3e27a46c5764b642a6eca663d1c6af752
    Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>

Feature: AudioComms
Category: Improvement
Purpose: AOSP+
Origin: Internal
Upstream-candidate: yes
Change-Id: Ibf8ca13193852398a6f6140324113c9d7aaefe84
Orig-Change-Id: I5571ed55dadce2b9751c9e2101a5d1cfca3e8479
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>