When setting a null device for an audio output, force the next routing request.
BZ: 53856
[CLEAN]

When a null device is set for an audio ouput, the outputDesc->mDevice is not updated,
thus the routing update is not done next time the device is selected for that output.
To meet the device change conditions, a boolean member mForceRouting has been
introduced to force the routing next time the output is set with a non-null device.

Change-Id: I27efdffcc3b577e26da1c759a2f674a64676618d
Orig-Change-Id: Ibf8ca13193852398a6f6140324113c9d7aaefe84
Category: device enablement
Domain: AudioComms-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>