UPSTREAM_PENDING: audio: add mechanism to re-route nullified audio
Fixes a race condition whereby two requests attempt to change the audio stream
(and interrupt each other).  For example, a user space app makes a call through
the audio API and before completion, a kernel function such as jack removal is
detected and is serviced.  In this case, you have a nullified audio route that
needs to be re-routed.

In order to know a routing update is needed, a boolean member mForceRouting
has been introduced to enforce routing for the next attempt to set audio output
with a non-null device.

Issue: GMIN-1567
Change-Id: I1c462621aa7fadf10c231fbff47feeb8f7523685
Upstream-URL: https://android-review.googlesource.com/99445
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Signed-off-by: Robert Beatty <robert.m.beatty@intel.com>