[PORT FROM MAIN][Workaround][All] JackDetection: No Audio when continuously plug/unplug wsHS (Audio routed to unpluged wsHS)
BZ: 56306

Whenever there is transition from IHF to headphone the mute count is incremented and
when the device change from headphone to headset the mute count is again incremented
before the mute count is decremented, therefore the value becomes 2. Now when device
is changed from headset to IHF the mute count is decremented only once. Hence the
mute will be still applied even after the wired headset is removed. This will lead to
no audio at the IHF. The fix will check whether the prev popcount is same as current
popcount value, if it is true then the mute count is not incremented.

Change-Id: Ife438ad6df4bd55918830bd2eb9d91cb9c0e05b5
orig-Change-Id: Id292d5b4abd768a132b8855ad8ff54c472732bb3
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>
Reviewed-on: http://android.intel.com:8080/82109
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>