[PORT FROM MAIN][AUDIO] Support new audio devices enum values introduced in JB MR1 code drop
BZ: 71815

Today, each audio device is identified by a bit in a 32 bit field with
16 bits for output devices and 16 bits for input devices.
So the number of output devices is now exhausted i.e. we can't add new output devices.

So in order to add new devices, a new convention for audio device
identification is defined, allowing 31 different input and output devices.
Bit 31 of the bit field is reserved and when set indicates an input device.

Orig-Change-Id: I66e28635e45505365c78733477b7bc669526ff28

Change-Id: I9b32706076fe86268cb1501ba35f517702fc3e15
Signed-off-by: ndifolcx <neilx.di.folco@intel.com>
Reviewed-on: http://android.intel.com:8080/123553
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>