[AUDIO] Stream Offload management with new route manager
BZ: 75313

Offload use cases were broken with introduction of new route
manager.
This patch changes the way policy and primary HAL (so Route manager)
communicate. It now use a setParameter with a StreamFlag key
when a compressed stream starts and stops.
It also still send the accessory change on the primary HAL when an
offload use case is on going.

Change-Id: I2cb45f4db84ddc1a2c3d418f8a556ac50c33467b
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/83482
Reviewed-by: Cahuzac, ThomasX <thomasx.cahuzac@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>