WiFi: rename intreface when does not match with prop.
BZ: 81693

It is needed to rename the interface from wpa_supplicant.conf
when it does not match with wifi.interface property.
The mechanism was not work if the new intreface name was
contained by the old one. example: new is wlan0 and
old wlan0:0.
This patch fix this issue by checking the length
of the two intrefaces to decide to update iface name.

Change-Id: I8394ab611ea20fe698b70aabba8dc370b0623487
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: http://android.intel.com:8080/87992
Reviewed-by: De Cesco, Jonathan <jonathan.de.cesco@intel.com>
Reviewed-by: Regairaz, Julie <julie.regairaz@intel.com>
Reviewed-by: Casasnovas, QuentinX <quentinx.casasnovas@intel.com>
Reviewed-by: Tricot, HonoreX <honorex.tricot@intel.com>
Reviewed-by: Mauger, MatthieuX <matthieux.mauger@intel.com>
Tested-by: Zurmely, PierreX <pierrex.zurmely@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>