[PORT FROM R42][AUDIO] When an input source is already used, give it to the new requester
BZ: 95092

When in communication with a connected headset/bluetooth device, the user
was forced to use that device's embedded mic even when he wanted to have
the output on speaker.

This fix allows selecting the built-in mic whenever the user wants to
force the speaker for output when in communication.

Change-Id: If9eb5cf7e151cb48ab662f44dc9a62ca124191c1
Origin-Change-Id: Iae8d66770262c11aad7e1ffb710a325bd608c759
Signed-off-by: pdiopX <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/101075
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>