Use built-in mic whenever IHF is forced as output device in call
BZ: 95092
[CLEAN]

When in communication with a connected headset/bluetooth device, the user
was forced to use that device's embedded mic even when he wanted to have
the output on speaker.

This fix allows selecting the built-in mic whenever the user wants to
force the speaker for output when in communication.

Change-Id: Iaa85171e505a7721e784e99f7c5b9694cb994422
Orig-Change-Id: I435b1572849f492c7f20920d42b9856933e39a0d
Category: device enablement
Domain: AudioComms-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: pdiopX <piotrx.diop@intel.com>