[PORT FROM MAIN][AUDIO] When an input source is already used, give it to the new requester
BZ: 95092

When in communication with a connected headset/bluetooth device, the user
was forced to use that device's embedded mic even when he wanted to have
the output on speaker.

This fix allows selecting the built-in mic whenever the user wants to
force the speaker for output when in communication.

Orig-Change-Id: If9eb5cf7e151cb48ab662f44dc9a62ca124191c1

Change-Id: I435b1572849f492c7f20920d42b9856933e39a0d
Signed-off-by: pdiopX <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/122710
Reviewed-by: Afonso, PhilippeX <philippex.afonso@intel.com>
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>