[FM][Audio] Implement support of force speaker in FM RX
BZ: 153948

Speaker is not supported during FM Rx playback,
as FM setParameter msg is not handled by the policy but only the audio hal.

This patch adds the support of speaker with respect of the policy.
MUSIC stream is used for FM Rx as the behaviour is the same than for media playback,
even if the stream is external.

Change-Id: I9956acc2fb91aa807fca4db37bd1c4b4b8a1a7bf
Category: device enablement
Domain: CWS.FM-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>