[REVERT ME][AUDIO]Change sync between compressed and primary on MRFLD
BZ: 89113

Sequence of open/close of the compressed stream and effective AUD
commands to route/unroute does not match the SAS/FAS sequences.

This patch restores the expected sequence documented in the SAS
for MRLD only but add a compile flag to avoid regression on CTP.
Need to clarify the behavior and remove this compile flag.

Change-Id: Ia0123368422c55d8ca87f05d473e7b49c17ee8e7
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: http://android.intel.com:8080/102301
Reviewed-by: Wagner, David <david.wagner@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>