[PORT FROM MAIN][ALL] JAVACRASH happened on app.odesanmi.and.music.fulleq when launch Music EQ app
BZ: 92162

In the AudioPolicyManagerBase there is condition in the registerEffect to check whether the
memory of effect has exceeded the max memory(512KB) of effects. With the new NXP effects
library the memory is exceeding this 512KB when multiple effects apps are launched. When the
memory exceeds this 512KB, AudioPolicyManagerBase is returning INVALID_OPERATION which is
causing crash in the application. This check is not required for Intel platforms, hence this
code is disabled.

Category: Feature Differentiation
Domain: AUDIO_MEDIA-Audio Policy
Origin: Internal
Upstream-Candidate: no, proprietary

Orig-Change-Id: I71dd971efc210c98860cfff118d5695676d00134
Orig-Change-Id: Ifa717b38d20e6ecfb6a439fb8e7257523a570167
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>
Signed-off-by: Arulselvan M <arulselvan.m@intel.com>

Change-Id: I8d39c1908409f6a251251f98c45aa7c914a85453
Signed-off-by: Arulselvan M <arulselvan.m@intel.com>