[PORT FROM R42][ALL] JAVACRASH happened on app.odesanmi.and.music.fulleq when launch Music EQ app
BZ: 92162

In the AudioPolicyManagerBase there is condition in the registerEffect to check whether the
memory of effect has exceeded the max memory(512KB) of effects. With the new NXP effects
library the memory is exceeding this 512KB when multiple effects apps are launched. When the
memory exceeds this 512KB, AudioPolicyManagerBase is returning INVALID_OPERATION which is
causing crash in the application. This check is not required for Intel platforms, hence this
code is disabled.

Change-Id: Ifa717b38d20e6ecfb6a439fb8e7257523a570167
Orig-Change-Id: Ib8b92bf9b00b4782b298d823568968919978462a
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>
Signed-off-by: Santosh <santoshx.joshi@intel>
Reviewed-on: http://android.intel.com:8080/105518
Reviewed-by: Joshi, SantoshX <santoshx.joshi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>