[PORT FROM MAIN][REMOTE BGM][DO NOT UPSTREAM] Local audio routes to remote sink when playing video only stream on remote sink
BZ: 101834

The audio policy routing to the remote sink gets initialized
only on first successful playback of an audio stream. In case
of video only stream, the audio path is not hit. So this
information is transmitted to HAL and retrieved in APM for
routing the audio to appropriate sinks.

Change-Id: I04ec0876f5b5bbe90716db7712bb8a60ae12954d
Orig-Change-Id: I966782a5693d4e196ccc3df2d7edd89c0de0207b
Signed-off-by: jerome <jerome.anand@intel.com>
Reviewed-on: http://android.intel.com:8080/122244
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>