[PORT FROM R4.1] Lost of 2 seconds of call when accepting it
BZ: 29679

The setMode in call was recently changed and is sent earlier.
The sequence is:
 - setMode In call from the policy
 - ModemCallStatus=true received from AT Manager
 - setStreamParameter with the device to use received from the policy.

As a result, when receiving the call status, the route is established
on the wrong device. When the setStreamParameter is received, a re-
routing is performed, increasing by around 800ms the establishment
of the audio call.

This patch ensures the policy always call setMode followed by
setStreamParameter on the output device.

Feature: AudioComms
Category: Integration
Purpose: PV-Enabler
Origin: Internal
Upstream-candidate: no
Change-Id: I44c774008f841e414e5025fba29ae1a0e24f7f3d
Orig-Change-Id: I596a4aa6971b570d352586060a67830dcd834167
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>