[PORT FROM MAIN][AUDIO] AudioPolicy: improve stopInput
BZ: 104921

Given the convention for audio device identification since JB MR1 code
drop, the bit 31 of the bit field is reserved and when set indicates
an input device. So 0x0 (AUDIO_DEVICE_NONE) is identified as an output
audio device. Thus the stopInput function does not stop the audio input
stream.

In order to reroute the audio input stream as expected, this patch
modifies the stopInput function in order to use the
AUDIO_DEVICE_BIT_IN (0x80000000) value.

Orig-Change-Id: I05fe4e3ac98cae2eb2cde951abd284b8b6540dbe

Change-Id: I04fc1bdb1b778c1ff5a9ab6edc6893a436fc5ea6
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Reviewed-on: http://android.intel.com:8080/122787
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>