[PORT FROM MAIN]During Music playback, choosing ringtones from settings play only on headset instead of headset+speaker.
BZ: 101884

The system stream and alarm stream are playing in the same output.
After stopping the output for touch tone APM is setting the device
for other active outputs which is playing music on HS.Check added
to verify there is no other streams playing on the same output as
the system stream. It will make sure that if there is alarm stream
playing on the same output APM will not set the device for other
active outputs.

Signed-off-by: sandhya <sandhyax.sebastian@intel.com>

Change-Id: I67e39e7aa6a7695bf91a9a22615eeeb67d12432e
Orig-Change-Id: Ibe5dd432c47e846339c2f853b3ccf7e8a29bf192
Signed-off-by: Arulselvan M <arulselvan.m@intel.com>
Reviewed-on: http://android.intel.com:8080/122263
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>