[PORT FROM MAIN][Audiocomms] Remove conditions that block the update of mForceUse in call
BZ: 112163

While in CSV call with an A2DP BT HS, which works in SCO mode in that UC,
if we plug a wHS, mForceUse will be set to FORCE_NO_BT_A2DP for MEDIA streams
as wHS as priority over A2DP for Media UCs.
But when it is unplugged, mForceUse should be set back to FORCE_NONE for
MEDIA streams, but it is blocked by some conditions.
Then when the call is ended, Music is played on IHF instead of playing on A2DP.

Fix consists in removing conditions that should not block the update of this
variable during CSV call.

Orig-Change-Id: I7fa0d935ce742bea008d97bdd16f0554afbcc8c6

Change-Id: I552ade7c59a5e39def2aafb7d2c11f5213ed4e01
Signed-off-by: ptonnetx <pierre-louisx.tonnet@intel.com>
Reviewed-on: http://android.intel.com:8080/122714
Reviewed-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>