[DO NOT UPSTREAM][AUDIO] Remove some BGM_ENABLED flags from AudioPolicyManagerBase
BZ: 114663

AudioPolicy volume attenuation of media stream in call is not functional.
This is due to an offset in mapping between AudioPolicyManagerBase
and AudioPolicyManagerALSA which are not compiled with the same flag
value; stack memory corruption occurs when initializing child members which
effectively share the same address as parent members. This corruption
makes the vector .indexOfKey() method fail.

This patch removes #ifdef BGM_ENABLED pre-compiler flags for background
music related properties of AudioPolicyManagerBase class. That way the
object size will be the same whatever the flag is set at.

ifdef BGM_ENABLED flags enclosing background music related methods declaration
are untouched to prevent future patches calling these methods when this flag
is set to FALSE on a particular product.

Change-Id: I84b0a5f93b6cbbce747d2bca677c311f87c1642f
Signed-off-by: Robin ZERBIB <robinx.zerbib@intel.com>