wifi-HAL: add a parallel HAL for bigcore platforms
BZ: 116022

The current WiFi HAL is closely tied to BCM which is expected.
It does loading of driver and also supports loading & reloading
of firmware.
The currently supported bigcore WiFi modules (atheros & iwlwifi)
do not need loading of driver nor does it need firmware update.
Also it does a rfkill start/stop for loading/unloading which is
sufficient enough

Since the functionalities are quite distinct we are using a
parallel HAL here

This HAL for bigcore has been ported from OTC tree which supports
bigcore platforms

Change-Id: Id05964a4a6733799ff23c792500bdf4f9be37e7c
Signed-off-by: Huzefa Kankroliwala <huzefa.nomanx.kankroliwala@intel.com>
Reviewed-on: http://android.intel.com:8080/113447
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Wood, Brian J <brian.j.wood@intel.com>
Tested-by: Ng, Cheon-woei <cheon-woei.ng@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>