[PORT FROM MAIN] wifi-HAL: add a parallel HAL for bigcore platforms
BZ: 116022

The current WiFi HAL is closely tied to BCM which is expected.
It does loading of driver and also supports loading & reloading
of firmware.
The currently supported bigcore WiFi modules (atheros & iwlwifi)
do not need loading of driver nor does it need firmware update.
Also it does a rfkill start/stop for loading/unloading which is
sufficient enough

Since the functionalities are quite distinct we are using a
parallel HAL here

This HAL for bigcore has been ported from OTC tree which supports
bigcore platforms

Change-Id: I1823909f559d731c8dee0a73fd34fdeaed85f002
Orig-Change-Id: Id05964a4a6733799ff23c792500bdf4f9be37e7c
Signed-off-by: Huzefa Kankroliwala <huzefa.nomanx.kankroliwala@intel.com>
Reviewed-on: http://android.intel.com:8080/123458
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Poulain, LoicX <loicx.poulain@intel.com>
Reviewed-by: Naoumenko, PaulX <paulx.naoumenko@intel.com>
Tested-by: Champciaux, Nicolas <nicolas.champciaux@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>