[PORT FROM MAIN][CTP-VB] [WiDi-BGM] When phone is routed to speaker phone the volume on background mode goes up.
BZ: 100971

When call is terminated, the volume of BGM output is updated based on "mBGMOutput" and device.
In the current code, after call termination while resuming BGM playback, before assigning the
"output" to "mBGMOutput", it calls checkAndSetVolume() and updates the wrong volume on BGM
output, as mBGMOutput = 0 and device is ear piece.
This fix checks "mIsBGMEnabled" before applying volume. Also the code is rearranged to avoid
unwanted volume changes on BGM output.
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>

Change-Id: I66c0549ece4e3fc758b98a6d890d286967414000
Orig-Change-Id: I78031b7b7e9c20311cd5f819320f15b8f7ffd047
Reviewed-on: http://android.intel.com:8080/122258
Reviewed-by: Bhakte, GurudattaX <gurudattax.bhakte@intel.com>
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>