[WiDi-BGM] Local audio-playback volume increases while Home button is selected on DUT and the background video-playback is active
BZ: 109056

In case of simultaneous playback of BGM and local music, when touch tones are
enabled, and when Home key is pressed, startOutput() is called to play touch
tone on Primary output and volume is updated for IHF, based on "volume index".
In the current code, the index is updated for WIDI as BGM is active but not for
IHF. So it takes default index to calculate volume for IHF, which causes this
issue. This fix uses WIDI volume index for IHF, when BGM is active.

Change-Id: Ia0bc71eb2035fc8d435311f2ac0d7a152d76ce8c
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>
Reviewed-on: http://android.intel.com:8080/116416
Reviewed-by: Krishna, Samaga <samaga.krishna@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Sebastian, SandhyaX <sandhyax.sebastian@intel.com>
Reviewed-by: Anand, Jerome <jerome.anand@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>