Make distinction between stopped and not-yet-started input
BZ: 115755
[CLEAN]

There was a race condition when a client calls getInput(), but is preempted
by a second client getInput() call for the same source. In that case, after
the release of the latter, the input is given back to the first client if he
is still alive (see patch I57b4b27151475d449d503d8e12d7d57159d04d77).
In this situation the startInput() call from the first client will fail as
the policy assumes that it has already started and is active.

This fix implements a new inputDescriptor parameter, mHasStarted to distinguish
an input that has been stopped (and therefore already called startInput()) from
an input that did not yet start.

Change-Id: I658122893cd57b7e9dceb87bd350d2ac005b98c7
Orig-Change-Id: Id4a57e8c13a5061338be3754075fbd5ae00ec5c3
Category: device enablement
Domain: AudioComms-Common
Origin: internal
Upstream-Candidate: no, needs rework
Signed-off-by: pdiopX <piotrx.diop@intel.com>