[Audio][DO NOT UPSTREAM] Reduce strategy unmute delay when switching outputs
BZ: 106108

While switching between devices eligible for different outputs, the policy
mutes the corresponding tracks to avoid glitches. During this operation, the
mute duration (MUTE_TIME_MS = 2s) was excessive and resulted in abnormal
delay in rerouting streams in few situations (e.g when inserting wsHS during
a playback in call).
This value was chosen arbitrarily high to ensure avoiding glitches, but intel
platforms possess a route manager that guarantees glitch-avoiding mechanics.

This fix reduces the unmute delay to the default value which corresponds to AF
mixer thread's buffer size revised upwards due to eventual delay in the audio
path, and gets the getDeviceForStrategy requests for STRATEGY_MEDIA to follow
STRATEGY_PHONE when in call.

Change-Id: I18bc5619c03bef85736603daa3398a1f10c7b560
Signed-off-by: Piotr Diop <piotrx.diop@intel.com>
Reviewed-on: http://android.intel.com:8080/120171
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Anand, Jerome <jerome.anand@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Krishna, Samaga <samaga.krishna@intel.com>
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Cahuzac, ThomasX <thomasx.cahuzac@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Tested-by: Joshi, SantoshX <santoshx.joshi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>