AudioPolicy: Add an "idle" strategy
BZ: 137361
[CLEAN]

This new strategy applies when no other does (hence the name "idle").
It elects a wired accessory if any is available or falls back to the
builtin speaker.

This allows the accessory change information to be forwarded even if there
is no active stream.  That is necessary for external routes that are not
represented by stream types (such as "media") in the audio policy.

Category: feature differentiation
Domain: AudioComms-VoiceTrigger
Origin: internal
Upstream-Candidate: yes
Change-Id: I415d1a5d113a66c00b4a168093bd77ec05520df3
Orig-Change-Id: I8c16485344b515a850b06e7c104cbba715354bba
Signed-off-by: David Wagner <david.wagner@intel.com>