AudioPolicy: Add an "idle" strategy
BZ: 137361

This new strategy applies when no other does (hence the name "idle").
It elects a wired accessory if any is available or falls back to the
builtin speaker.

This allows the accessory change information to be forwarded even if there
is no active stream.  That is necessary for external routes that are not
represented by stream types (such as "media") in the audio policy.

Change-Id: I8c16485344b515a850b06e7c104cbba715354bba
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: http://android.intel.com:8080/124207
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>