[Music Offload] Next song is not played in playlist
BZ: 130001

This patch fixes the issue of playing multiple songs in loop
in a playlist. This also holds good for gapless songs. The
code will check the reference count for a stream in
releaseOutput function of APM. This was the breakage caused
while porting to MR2.

Technically this patch:
   The stream type need to be considered when reading the
   reference count of the output descriptor.
   In offload case the stream is alwasy MUSIC.

Category: feature differentiation
Domain: audio-policy
Severity: High
Origin: internal
upstream-candidate: no, proprietary

Change-Id: I66b3315de77eba3a6b832bb96e18ce2bf3a102e8
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>
Reviewed-on: http://android.intel.com:8080/128461
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>