Music is played from loudspeaker not from wired headset after music playback is interrupted by alarm clock.
BZ: 107199

After the alarm clock expired the device selection for wsHS was not being done
in stopout for the deep buffer stream if the alarm tone was offloaded.

After stopping the direct output the fix will check for deep buffer flag and
primary output also and will select wsHS for the playback of background music.
Also the code will check the current output is active or not.If it is not active
then only policymanager will check for other active outputs.This was a breakage
caused while porting to MR2.

Category: feature differentiation
Domain: Audio Media-Audio Policy
Severity: medium
Origin: internal
Upstream-Candidate:  no, proprietary

Change-Id: I616d5807c042e01887c635b93b6b77b628e08e71
Signed-off-by: sandhya <sandhyax.sebastian@intel.com>
Reviewed-on: http://android.intel.com:8080/129503
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Sameullah, MazharX <mazharx.sameullah@intel.com>
Reviewed-by: Anand, Jerome <jerome.anand@intel.com>
Reviewed-by: Choudhari, SadiqX <sadiqx.choudhari@intel.com>
Reviewed-by: Krishna, Samaga <samaga.krishna@intel.com>
Tested-by: Gupta, ArvindX K <arvindx.k.gupta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>