[Workaround] Music is played from IHF for a sec after alarm tone is alerted when music is being played on WsHS.
BZ: 106491

This is a workaround, ideal fix should be querying the HAL for deepbuffer latency
which returns based on the buffer size configured.

Deep buffer has a larger buffer of size 192ms as compared to normal bufeer of
96ms, and hence the latency of deepbuffer is double that of normal buffer.

This patch updates the "muteWaitMs" to correct value for deep buffer, so that
the PCM buffers are completely drained before proceeding.

Category: Feature differentiation
Domain: AUDIO_MEDIA, Audio Policy
Origin: Internal
Upstream-Candidatet: No, proprietary

Change-Id: I89669eb02f88260f407ab1f623210bb8719b7236
Signed-off-by: guruprasadX, pawse <guruprasadx.pawse@intel.com>