[AUDIO] configure setParameters for LPAL in policy
BZ: 156552

This patch implements the support for Low-Power
Always Listening (LPAL) audio source.
The audio source LPAL is low priority and do not
contain a stream in upper layers.
So the policy has to set the parameters to
vstv-route=on and the lpal_device to the correct value.
This device value should be converted to the new API
revision 2.0.

Change-Id: I0529a66f693cb70329c141bbc7db0ba565b8cb17
Category: feature differentiation
Domain: AudioComms-VoiceTrigger
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Cedric Chapuis <cedricx.chapuis@intel.com>
Signed-off-by: pafonsoX <philippex.afonso@intel.com>