Crash in checkAndSetVolume function while playing only video file in remote background mode and playing another video file locally
BZ: 153794

The issue occurs when only video file is played in the remote
background mode and when you try to play another video file locally.
BGM is applicable only for music streams and there is a possibility
of inactive output getting returned based on a different strategy
and try to apply volume for this output. The fix will make sure that
the volume is not applied for this output

Change-Id: I1c66fb77f125fd5d4cff62282c9e408ee080daeb
Category: device enablement
Domain: AUDIO_MEDIA-Audio Policy
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>