[Deep Buffer] Playing music , audio burst after plug out wsHS and resume the song instantaneously
BZ: 159694

For deep buffer playback the buffersize is 96ms.When plugin/plugout
wsHS the device switch will happen before draining out the prepared
data.If the touch tone is enabled the device switch will happen for
the primary output and only primary output is muted and deep buffer
output will not be muted.So the prepared data will be played out on
the new device.

If the primary output is active when the device switch happens muting
both the primary and deep buffer outputs so that the audio will be muted
and there will not be any audio burst.

Change-Id: I393ce22f84fd9ba352d8ecffa001f380553165ef
Category: device enablement
Domain: AUDIO_MEDIA-Audio Policy
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: sandhya <sandhyax.sebastian@intel.com>