audio: close used inputs before opening a new one
The audio system expects routing to be done in HAL. The
problem is that we have a separate HAL for bluetooth hsp
and we cannot connect the intel HSB audio HAL with
the BT hsp HAL. We overcome this issue by invalidating
the tracks when we switch from a BT device to a nonBT device
or from a nonBT device to a BT device. When we do this
the startInput method is called and we have to close the
used inputs before opening a new one.

For: APDEV-891
Category: AOSP Improvement
Domain: AOSP-audio
Origin: internal
Upstream-Candidate: No
Depends-Change-Id: Ibadecf774925262d45027f16bbe6ee61224f285c
Change-Id: Id93c6c3723dc433d4bc1dee4f6123d9bcb37565d
Signed-off-by: Constantin Musca <constantin.musca@intel.com>
Reviewed-on: https://android.intel.com/175333
Reviewed-by: Mathew, AlwinX <alwinx.mathew@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Zhi, Yong <yong.zhi@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>