Video loses audio when pressing control button on wired headset to resume/pause music.
BZ: 161402

While video is playing in the background and if we try to resume
the mp3 file through the wired headset control button, the mp3 file
continues to get offloaded and audio output is mixed. When we try
to pause the mp3 file, setparameter is invoked in stopoutput for the
offload case to inform primary HAL that offload output is stopped,
but the route manager closes playback route of all outputs, which is
causing this issue. For resolving this issue, setParameter is invoked
only for offload output.

Change-Id: Ie893dbba580f694c815f34b0a14d3593333644fc
Category: feature differentiation
Domain: AUDIO_MEDIA-LPA
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>