[AUDIO] Set isVirtualInputDevice method as protected
BZ: 174189

isVirtualInputDevice is needed by the derived audio policy,
to check if the current input device is a virtual device.
This methods is currently private to the base policy class,
and can not be called from the daugther policy class.

Set protected access to the isVirtualInputDevice method, so
that it can be called from the derived policy.

Change-Id: I623d4b88ce0de13092b949eff7814d161ecd85eb
Category: device enablement
Domain: AUDIO_MEDIA-WIDI
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>