Screen display stuck 1s and the video playback time will move back/forward a few seconds when plugging in wsHS.
BZ: 179437

When wsHS is plugged in, Offload device is opened and closed. But after offload
output is opened, in checkOutputForStrategy() due to mismatch in vectors
srcOutputs and dstOutputs, setStreamOutput() is invoked. This causes AudioFlinger
to invalidate the existing Track, though there is no chnage in output, which
leads to creation of new Track and Tear Down, and finally this issue.
This patch fixes this issue by allowing checkOutputForStrategy() to move the
strategy from one output to another only if the outputs are different.

Change-Id: I5bd786e8088ee381c96bf616f94cfc15e656010d
Category: aosp improvement
Domain: AUDIO_MEDIA-Audio Policy
Origin: internal
Upstream-Candidate: yes
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>