WiFi: Update wpa_supplicant.conf only if needed
BZ: 198381

An older patch has been commited in order to append the
wpa_supplicant.conf with a "ctrl_interface" field.
However, this first patch adds systematicaly this field
at each WiFi power ON, even if it is already present.
This leads to undesired file write and finally the file
can contain hundreds of "ctrl_interface" iterrations.
Reference of that patch:
    wifi: fix scalable HAL for supplicant
    Change id: I266cf940630106161a06267e64adefdf702294f7
    Author: Jeremie GARCIA <jeremie.garcia@intel.com>
    Date: Nov 29, 2013

Current patch checks first if "ctrl_interface" field is
present before deciding to add it.

In addition, in case of config file is created from the
template, an enhancement has been done to be sure that
the file is really closed before reopening it in append
mode.

Change-Id: Ic9dceef6e2e9bc2902e9a2c3f64f466e633fbe47
Category: aosp improvement
Domain: CWS.WIFI-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Matthieu MAUGER <matthieux.mauger@intel.com>