audio: close used inputs before opening a new one
The audio system expects routing to be done in HAL. The
problem is that we have a separate HAL for bluetooth hsp
and we cannot connect the intel HSB audio HAL with
the BT hsp HAL. We overcome this issue by invalidating
the tracks when we switch from a BT device to a nonBT device
or from a nonBT device to a BT device. When we do this
the startInput method is called and we have to close the
used inputs before opening a new one.

For: APDEV-891
Category: AOSP Improvement
Domain: AOSP-audio
Origin: internal
Upstream-Candidate: No

Change-Id: Ie9032319d8718ec2fec6d4f9694af00ae8befdbd
Signed-off-by: Constantin Musca <constantin.musca@intel.com>
Reviewed-on: https://android.intel.com/175333
Reviewed-by: Mathew, AlwinX <alwinx.mathew@intel.com>
Tested-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Zhi, Yong <yong.zhi@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Reviewed-on: https://android.intel.com/299550
Reviewed-by: Rajendran, SadhanaX <sadhanax.rajendran@intel.com>
Tested-by: Rajendran, SadhanaX <sadhanax.rajendran@intel.com>
Reviewed-by: Sarasetti, ManteshX <manteshx.sarasetti@intel.com>
Reviewed-by: Vinjamuri, Venkateswarlu V <venkateswarlu.v.vinjamuri@intel.com>
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>